Add account. account type, role, user role

+++ SanyaaDelivery.Domain/Models/FawryChargeT.cs
@@ -11,7 +11,7 @@ namespace SanyaaDelivery.Domain.Models
         }
 
         public int SystemId { get; set; }
-        public int? FawryRefNumber { get; set; }
+        public long? FawryRefNumber { get; set; }
         public string ChargeStatus { get; set; }
         public double? ChargeAmount { get; set; }
         public DateTime? ChargeExpireDate { get; set; }
